Block GdCell::borrow and Borrow_mut on other threads
